- Changed the project design to fit the requirements

+++ src/main/java/itsix/CreditProject/models/IInterval.java
@@ -2,8 +2,8 @@ package itsix.CreditProject.models;
 
 public interface IInterval {
 
-	Double getMin();
+	Integer getMin();
 
-	Double getMax();
+	Integer getMax();
 
 }
